requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amazon-dax-client==1.1.3
antlr4-python3-runtime==4.7.2  # via amazon-dax-client
botocore==1.13.46         # via amazon-dax-client
click==7.0                # via pip-tools
django==2.2.4
docutils==0.15.2          # via botocore
jmespath==0.9.4           # via botocore
mod_wsgi==4.6.7
pip-tools==4.2.0
python-dateutil==2.8.1    # via botocore
pytz==2019.3              # via django
six==1.13.0               # via amazon-dax-client, pip-tools, python-dateutil
sqlparse==0.3.0           # via django
urllib3==1.25.7           # via botocore
